Refactor main so that it immediately switches on arg0
Put the different modes of operation each in their own module.
